+++ dg-muscle-ios/sources/Test/History/Test/GroupByMonthHistoriesUsecaseTests.swift
@@ -7,12 +7,15 @@
 
 import Foundation
 import XCTest
+import MockData
 import Domain
 
 final class GroupByMonthHistoriesUsecaseTests: XCTestCase {
     func test() {
         let usecase = GroupByMonthHistoriesUsecase()
-        let groupedByMonthHistories = usecase.implement(histories: HistoryRepository()._histories)
+        let groupedByMonthHistories = usecase.implement(histories: [
+            HISTORY_1, HISTORY_2, HISTORY_3
+        ])
         
         XCTAssertEqual(groupedByMonthHistories["202405"]?.count, 2)
         XCTAssertEqual(groupedByMonthHistories["202404"]?.count, 1)
